Separate DoubleClick RTC tests into their own file (#10747)
* start refactor

* continue refactor

* All RTC tests passing here

* Finish refactor

* Change name of test block

* Fix lint

* Removed fixtures, responded to feedback

* fix presubmit issues

* Fix jsdoc types
